refactor plugins into separate files for better modularity
