+++ .config/nvim/lua/plugins/gitsigns.lua
@@ -1,0 +1,35 @@
+-- show changes in file, when working in a git repo
+return {
+  {
+    "lewis6991/gitsigns.nvim",
+    dependencies = {
+      "nvim-lua/plenary.nvim",
+    },
+    opts = {
+      on_attach = function(buffer)
+        local gs = package.loaded.gitsigns
+
+        local function map(mode, l, r, desc)
+          vim.keymap.set(mode, l, r, { buffer = buffer, desc = desc })
+        end
+
+        map("n", "]h", gs.next_hunk, "Next hunk")
+        map("n", "[h", gs.prev_hunk, "Prev hunk")
+        map({ "n", "v" }, "<leader>hs", ":Gitsigns stage_hunk<CR>", "Stage hunk")
+        map({ "n", "v" }, "<leader>hr", ":Gitsigns reset_hunk<CR>", "Reset hunk")
+        map("n", "<leader>hS", gs.stage_buffer, "Stage buffer")
+        map("n", "<leader>hu", gs.undo_stage_hunk, "Undo Stage hunk")
+        map("n", "<leader>hR", gs.reset_buffer, "Reset buffer")
+        map("n", "<leader>hp", gs.preview_hunk, "Preview hunk")
+        map("n", "<leader>hb", function()
+          gs.blame_line({ full = true })
+        end, "Blame line")
+        map("n", "<leader>hd", gs.diffthis, "Diff this")
+        map("n", "<leader>hD", function()
+          gs.diffthis("~")
+        end, "Diff This ~")
+        map({ "o", "x" }, "ih", ":<C-U>Gitsigns select_hunk<CR>", "GitSigns Select hunk")
+      end,
+    },
+  },
+}
