+++ .config/nvim/lua/plugins/mason.lua
@@ -1,0 +1,8 @@
+-- convenient installation of LSP clients
+return {
+  {
+    "williamboman/mason.nvim",
+    lazy = false,
+    config = true,
+  },
+}
