+++ .config/nvim/lua/plugins/stickybuf.lua
@@ -1,0 +1,11 @@
+-- Prevents opening buffers in terminal buffer
+return {
+  {
+    "stevearc/stickybuf.nvim",
+    opts = {
+      get_auto_pin = function(bufnr)
+        return require("stickybuf").should_auto_pin(bufnr)
+      end,
+    },
+  },
+}
